% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AIC.jmodelMult.R
\name{AIC.jmodelMult}
\alias{AIC.jmodelMult}
\title{Return the AIC score associated with a jmodelMult object (a joint model with NMRE)}
\usage{
\method{AIC}{jmodelMult}(object, ...)
}
\arguments{
\item{object}{A jmodelMult object as this is produced by function jmodelMult}

\item{...}{Not used.}
}
\description{
Return the AIC score associated with a jmodelMult object (a joint model with NMRE)
}

