% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AIC.jmodelTM.R
\name{AIC.jmodelTM}
\alias{AIC.jmodelTM}
\title{Return the AIC score associated with a jmodelTM object}
\usage{
\method{AIC}{jmodelTM}(object, ...)
}
\arguments{
\item{object}{A jmodelTM object as this is produced by function jmodelTM}

\item{...}{Not used.}
}
\description{
Return the AIC score associated with a jmodelTM object
}

