% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BIC.jmodelMult.R
\name{BIC.jmodelMult}
\alias{BIC.jmodelMult}
\title{Return the BIC score associated with a jmodelMult object (a joint model with NMRE)}
\usage{
\method{BIC}{jmodelMult}(object, ...)
}
\arguments{
\item{object}{A jmodelMult object as this is produced by function jmodelMult}

\item{...}{Not used.}
}
\description{
Return the BIC score associated with a jmodelMult object (a joint model with NMRE)
}

