% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BIC.jmodelTM.R
\name{BIC.jmodelTM}
\alias{BIC.jmodelTM}
\title{Return the BIC score associated with a jmodelTM object}
\usage{
\method{BIC}{jmodelTM}(object, ...)
}
\arguments{
\item{object}{A jmodelTM object as this is produced by function jmodelTM}

\item{...}{Not used.}
}
\description{
Return the BIC score associated with a jmodelTM object
}

