% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CheckInputs.R
\name{CheckInputs}
\alias{CheckInputs}
\title{Check that the basic input are reasonable}
\usage{
CheckInputs(fitLME, fitCOX, rho)
}
\arguments{
\item{fitLME}{: fitted Linear Mixed Effects model}

\item{fitCOX}{: fitted Proportional Hazards Regression model}

\item{rho}{:  number specifying the logarthmic transformation model}
}
\value{
VOID
}
\description{
Check that the basic input are reasonable
}

