% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GenerateControlList.R
\name{GenerateControlList}
\alias{GenerateControlList}
\title{Generate the control variables used by the model}
\usage{
GenerateControlList(control)
}
\arguments{
\item{tol.P}{: tolerance for parameters (numeric)}

\item{tol.L}{: tolerance for log-likelihood  (numeric)}

\item{max.iter}{: maximum number of EM iterations (numeric)}

\item{SE.method}{: standard error estimation method  (char)}

\item{delta}{: increment used by forward difference  (numeric)}

\item{nknot}{: number of Gauss-Hermite quadrature knots  (numeric)}
}
\value{
controlVals list
}
\description{
Generate the control variables used by the model
}

