% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jmodelMult.R
\name{jmodelMult}
\alias{jmodelMult}
\title{Joint Modeling Main Function with NMRE (nonparametric Multiplicative random effects)}
\usage{
jmodelMult(fitLME, fitCOX, data, model = 1, rho = 0, timeVarT = NULL,
  control = list(), ...)
}
\arguments{
\item{fitLME}{is a}

\item{fitCOX}{is a}

\item{data}{is a}

\item{model}{is a}

\item{rho}{is a}

\item{timeVarT}{is a}

\item{control}{is a}

\item{...}{can be additional arguments}
}
\value{
jmodelMult class model
}
\description{
Joint Modeling Main Function with NMRE (nonparametric Multiplicative random effects)
}
\examples{
1 + 3
}

