% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jmodelTM.R
\name{jmodelTM}
\alias{jmodelTM}
\title{Joint Modeling Main Function with LME (linear mixed effects)}
\usage{
jmodelTM(fitLME, fitCOX, data, model = 1, rho = 0, timeVarY = NULL,
  timeVarT = NULL, control = list(), ...)
}
\arguments{
\item{fitLME}{is a}

\item{fitCOX}{is a}

\item{data}{is a}

\item{model}{is a}

\item{rho}{is a}

\item{timeVarY}{is a}

\item{timeVarT}{is a}

\item{control}{is a}

\item{...}{can be additional arguments}
}
\value{
jmodelTM class model
}
\description{
Joint Modeling Main Function with LME (linear mixed effects)
}
\examples{
1 + 3
}

