% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logLik.jmodelMult.R
\name{logLik.jmodelMult}
\alias{logLik.jmodelMult}
\title{Print the logLikelihood of joint model model with NMRE}
\usage{
\method{logLik}{jmodelMult}(object, ...)
}
\arguments{
\item{object}{A jmodelMult object}

\item{...}{Not used.}
}
\description{
Print the logLikelihood of joint model model with NMRE
}

