% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logLik.jmodelTM.R
\name{logLik.jmodelTM}
\alias{logLik.jmodelTM}
\title{Print the logLikelihood of joint model model}
\usage{
\method{logLik}{jmodelTM}(object, ...)
}
\arguments{
\item{object}{A jmodelTM object.}

\item{...}{Not used.}
}
\description{
Print the logLikelihood of joint model model
}

