% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.jmodelMult.R
\name{print.jmodelMult}
\alias{print.jmodelMult}
\title{Print basic summary of joint model model with NMRE}
\usage{
\method{print}{jmodelMult}(x, digits = max(4, getOption("digits") - 4), ...)
}
\arguments{
\item{x}{A jmodelMult object}

\item{digits}{A scalar defining how many digits to use when printing out the summary (min:4)}

\item{...}{Not used.}
}
\description{
Print basic summary of joint model model with NMRE
}

