% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.jmodelTM.R
\name{print.jmodelTM}
\alias{print.jmodelTM}
\title{Print basic summary of joint model model}
\usage{
\method{print}{jmodelTM}(x, digits = max(4, getOption("digits") - 4), ...)
}
\arguments{
\item{x}{A jmodelTM object.}

\item{digits}{A scalar defining how many digits to use when printing out the summary (min:4)}

\item{...}{Not used.}
}
\description{
Print basic summary of joint model model
}

