% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.summary.jmodelMult.R
\name{print.summary.jmodelMult}
\alias{print.summary.jmodelMult}
\title{Print detailed summary of joint model model with NMRE}
\usage{
\method{print}{summary.jmodelMult}(x, digits = max(4, getOption("digits") -
  4), ...)
}
\arguments{
\item{x}{A jmodelMult object.}

\item{digits}{A scalar defining how many digits to use when printing out the summary (default & min: 4)}

\item{...}{Not used.}
}
\description{
Print detailed summary of joint model model with NMRE
}

