% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.summary.jmodelTM.R
\name{print.summary.jmodelTM}
\alias{print.summary.jmodelTM}
\title{Print detailed summary of joint model model}
\usage{
\method{print}{summary.jmodelTM}(x, digits = max(4, getOption("digits") - 4),
  ...)
}
\arguments{
\item{x}{A jmodelTM object.}

\item{digits}{A scalar defining how many digits to use when printing out the summary (default & min: 4)}

\item{...}{Not used.}
}
\description{
Print detailed summary of joint model model
}

