% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.jmodelMult.R
\name{summary.jmodelMult}
\alias{summary.jmodelMult}
\title{Compute summaries of objects in the joint model class with NMRE}
\usage{
\method{summary}{jmodelMult}(object, ...)
}
\arguments{
\item{object}{A jmodelMult object as this is produced by function jmodelMult}

\item{...}{Not used.}
}
\description{
Compute summaries of objects in the joint model class with NMRE
}

