% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.jmodelTM.R
\name{summary.jmodelTM}
\alias{summary.jmodelTM}
\title{Compute summaries of objects in the joint model class}
\usage{
\method{summary}{jmodelTM}(object, ...)
}
\arguments{
\item{object}{A jmodelMult object as this is produced by function jmodelTM}

\item{...}{Not used.}
}
\description{
Compute summaries of objects in the joint model class
}

