% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vcov.jmodelMult.R
\name{vcov.jmodelMult}
\alias{vcov.jmodelMult}
\title{Return the variance covariance matrix of a jmodelMult object (a joint model with NMRE)}
\usage{
\method{vcov}{jmodelMult}(object, ...)
}
\arguments{
\item{object}{A jmodelMult object as this is produced by function jmodelMult}

\item{...}{Not used.}
}
\description{
Return the variance covariance matrix of a jmodelMult object (a joint model with NMRE)
}

