% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vcov.jmodelTM.R
\name{vcov.jmodelTM}
\alias{vcov.jmodelTM}
\title{Return the variance covariance matrix of a jmodelTM object}
\usage{
\method{vcov}{jmodelTM}(object, ...)
}
\arguments{
\item{object}{A jmodelTM object as this is produced by function jmodelTM}

\item{...}{Not used.}
}
\description{
Return the variance covariance matrix of a jmodelTM object
}

